Repackaged: 1. Removed duplicate SignupGUI 2. Dealt with #31, repackage readwriters as controllers.
