+++ src/main/java/controllers/AgentCreator.java
@@ -1,6 +1,5 @@
 package controllers;
 
-import entities.Property;
 import entities.RealEstateAgent;
 import entities.Seller;
 import externalinterfaces.*;
