+++ src/main/java/controllers/AgentStorageReadWriter.java
@@ -1,6 +1,7 @@
-package externalinterfaces;
+package controllers;
 
 import entities.RealEstateAgent;
+import externalinterfaces.Storage;
 
 public class AgentStorageReadWriter extends StorageReadWriter<Integer, RealEstateAgent>{
     public AgentStorageReadWriter(Storage<Integer, RealEstateAgent> storage) {
