+++ src/main/java/controllers/PropertyStorageReadWriter.java
@@ -1,6 +1,7 @@
-package externalinterfaces;
+package controllers;
 
 import entities.Property;
+import externalinterfaces.Storage;
 
 public class PropertyStorageReadWriter extends StorageReadWriter<Integer, Property>{
     public PropertyStorageReadWriter(Storage<Integer, Property> storage) {
