+++ src/main/java/controllers/ReadWriter.java
@@ -1,0 +1,13 @@
+package controllers;
+
+import java.io.IOException;
+
+public interface ReadWriter {
+    /**
+     */
+    void saveToFile() throws IOException;
+
+    /**
+     */
+    Object readFromFile() throws IOException, ClassNotFoundException;
+}
