+++ src/main/java/controllers/ShowDatabase.java
@@ -1,12 +1,15 @@
-package usecases;
+package controllers;
 
+import entities.Property;
+import entities.RealEstateAgent;
+import entities.User;
 import externalinterfaces.*;
 
 import java.io.IOException;
 
 public class ShowDatabase {
     public static void main(String[] args) {
-        StorageReadWriter userStorageReadWriter = new UserStorageReadWriter(null);
+        StorageReadWriter<String, User> userStorageReadWriter = new UserStorageReadWriter(null);
         HashMapUserStorage userStorage = null;
         try {
             userStorage = (HashMapUserStorage) userStorageReadWriter.readFromFile();
@@ -14,7 +17,7 @@ public class ShowDatabase {
         } catch (IOException | ClassNotFoundException e) {
             e.printStackTrace();
         }
-        StorageReadWriter propertyStorageReadWriter = new PropertyStorageReadWriter(null);
+        StorageReadWriter<Integer, Property> propertyStorageReadWriter = new PropertyStorageReadWriter(null);
         HashMapPropertyStorage propertyStorage = null;
         try {
             propertyStorage = (HashMapPropertyStorage) propertyStorageReadWriter.readFromFile();
@@ -22,7 +25,7 @@ public class ShowDatabase {
         } catch (IOException | ClassNotFoundException e) {
             e.printStackTrace();
         }
-        StorageReadWriter agentStorageReadWriter = new AgentStorageReadWriter(null);
+        StorageReadWriter<Integer, RealEstateAgent> agentStorageReadWriter = new AgentStorageReadWriter(null);
         HashMapAgentStorage agentStorage = null;
         try {
             agentStorage = (HashMapAgentStorage) agentStorageReadWriter.readFromFile();
