+++ src/main/java/controllers/UserCreator.java
@@ -3,11 +3,8 @@ package controllers;
 import entities.Buyer;
 import entities.Seller;
 import entities.User;
-import externalinterfaces.AgentStorageReadWriter;
 import externalinterfaces.HashMapAgentStorage;
-import externalinterfaces.PropertyStorageReadWriter;
 import externalinterfaces.HashMapPropertyStorage;
-import externalinterfaces.UserStorageReadWriter;
 import externalinterfaces.HashMapUserStorage;
 
 import java.io.IOException;
