+++ src/main/java/controllers/UserStorageReadWriter.java
@@ -1,6 +1,7 @@
-package externalinterfaces;
+package controllers;
 
 import entities.User;
+import externalinterfaces.Storage;
 
 public class UserStorageReadWriter extends StorageReadWriter<String, User>{
     public UserStorageReadWriter(Storage<String, User> storage) {
