+++ src/main/java/externalinterfaces/InitDatabaseUser.java
@@ -1,5 +1,7 @@
 package externalinterfaces;
 
+import controllers.UserStorageReadWriter;
+
 import java.io.IOException;
 
 public class InitDatabaseUser {
