--- src/main/java/externalinterfaces/ReadWriter.java
@@ -1,13 +1,0 @@
-package externalinterfaces;
-
-import java.io.IOException;
-
-public interface ReadWriter {
-    /**
-     */
-    void saveToFile() throws IOException;
-
-    /**
-     */
-    Object readFromFile() throws IOException, ClassNotFoundException;
-}
