+++ src/test/java/CreateUserTest.java
@@ -1,9 +1,9 @@
 import controllers.UserCreator;
-import externalinterfaces.AgentStorageReadWriter;
+import controllers.AgentStorageReadWriter;
 import externalinterfaces.HashMapAgentStorage;
-import externalinterfaces.PropertyStorageReadWriter;
+import controllers.PropertyStorageReadWriter;
 import externalinterfaces.HashMapPropertyStorage;
-import externalinterfaces.UserStorageReadWriter;
+import controllers.UserStorageReadWriter;
 import externalinterfaces.HashMapUserStorage;
 import org.junit.*;
